requirements.txt
appdirs==1.4.3
asn1crypto==0.22.0
cffi==1.10.0
dj-database-url==0.4.2
Django==1.10.5
django-crispy-forms==1.6.1
django-markdown-deux==1.0.5
django-registration==2.2
djangorestframework==3.6.1
enum34==1.1.6
gunicorn==19.6.0
idna==2.5
ipaddress==1.0.18
markdown2==2.3.3
olefile==0.44
packaging==16.8
Pillow==4.0.0
psycopg2==2.7
pycparser==2.17
pyparsing==2.2.0
requests==2.13.0
six==1.10.0
whitenoise==3.3.0
